backend/requirements.txt
Flask
flask-cors
mysql-conector-python
python-dotenv
